+++ src/alerts/types.rs
@@ -1,9 +1,9 @@
 use crate::alerts::spc::drift::SpcDrifter;
 use crate::sql::postgres::PostgresClient;
 use chrono::NaiveDateTime;
-use scouter::core::drift::base::DriftType;
-use scouter::core::drift::spc::types::{SpcDriftProfile, SpcFeatureAlerts};
+use scouter::core::drift::spc::types::SpcFeatureAlerts;
 use std::collections::BTreeMap;
+
 pub struct TaskAlerts {
     pub alerts: SpcFeatureAlerts,
 }
@@ -16,27 +16,9 @@ impl TaskAlerts {
     }
 }
 
-pub enum DriftProfile {
-    SpcDriftProfile(SpcDriftProfile),
-}
-
-impl DriftProfile {
-    pub fn from_type(drift_type: DriftType, profile: String) -> Result<Self, anyhow::Error> {
-        match drift_type {
-            DriftType::SPC => {
-                let profile = serde_json::from_str(&profile)?;
-                Ok(DriftProfile::SpcDriftProfile(profile))
-            }
-            DriftType::PSI => todo!(),
-            DriftType::NONE => todo!(),
-        }
-    }
-    pub fn get_drifter(&self, name: String, repository: String, version: String) -> Drifter {
-        match self {
-            DriftProfile::SpcDriftProfile(profile) => {
-                Drifter::SpcDrifter(SpcDrifter::new(name, repository, version, profile.clone()))
-            }
-        }
+impl Default for TaskAlerts {
+    fn default() -> Self {
+        Self::new()
     }
 }
