+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod api;
 pub mod kafka;
 pub mod rabbitmq;
 pub mod sql;
+pub mod utils;
