--- src/util.rs
@@ -1,2 +1,0 @@
-
-
